edited total type as float

+++ day_2_project.py
@@ -1,5 +1,5 @@
 print("Welcome to the tip calculator!")
-total = int(input("What was the total bill? $"))
+total = float(input("What was the total bill? $"))
 percent = int(input("How much tip would you like to give? 10, 12, or 15?"))
 split = int(input("How many people to split the bill?"))
 tip = total * (percent / 100)
